working leader election netween esp and osx

+++ lib/raft/src/logEntry.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct LogEntry {
-  unsigned int term;
+  unsigned short term;
   unsigned char type;
   unsigned char data[8];
 };
